Modified nodes to send orientation data (currently simulated movements).

+++ biot-applications/modules/sendData/sendData.h
@@ -3,10 +3,10 @@
 
 typedef struct nodeDataStructure {
     uint32_t timeStamp;
-    int16_t w;    
-    int16_t x;    
-    int16_t y;    
-    int16_t z;    
+    float w;    
+    float x;    
+    float y;    
+    float z;    
 } nodeData_t;
 
 void sendData(char *address, nodeData_t data);
